TRAP? => ERROR? TRAP, kill CATCH?
A change in the protocol for TRAP means it is now possible to tell
"from the outside" if it trapped an error or not.  This is because
if it ever returns an ERROR!, it means that error was actually raised.
Other errors are turned into void...and if this is a problem, then
one needs to use TRAP/WITH to separate out the handling of errors
that were evaluated to normally vs raised.

Yet typically, processing of a TRAP is not of a routine that could
return an ERROR! (unless one is writing truly generic code, like a
console).  So ERROR? TRAP is a good enough test and is clearer.

The CATCH? function doesn't have any such discernment, but it really
seems kind of useless to test for "was there a catch".  So since this
whole family of refinements is being removed, it is too.

Simplifies TRAP to take advantage of the new automatic releasing of
handles when there is a fail() on their frame stack.  Also cuts down
on excessive commenting in TRAP.

Also fixes a bug by which the scanner was not putting newline info
on PATH! values.