Fix pair recycling, edge condition in stack assert
Which part of a "pairing" series was considered the REBSER node changed
at one point, but the GC marking for REB_PAIR didn't take that into
account...so it was marking the second value in the pairing.  Hence
pairs that were still in use could be recycled.  This fixes that.

Also changes DS_AT() so that it tolerates indexing into the stack at
one past the top, to use as a limit for copying.  Previously, it was
possible to potentially see an END marker at one past the top in edge
cases and assert.