Patch ELIDE bug in REDUCE, assert for related bugs
REDUCE reused a frame without explicitly setting the DO_FLAG_XXX, which
meant that if something like ELIDE were to set DO_FLAG_NO_LOOKAHEAD
as part of its operation then it would not get cleared on the next
call to Do_Core().  The consequence was that code like:

    reduce [1 + 2 elide (3 + 4) 5 + 6]

...would not work, because the ELIDE would shift the frame into a no
lookahead mode that would not be cleared, so the 5 after the ELIDE
would evaluate without picking up the +, leaving (+ 6) to run and
cause an error.

This addresses that issue by saving and restoring the frame flags on
each Do_Next that reuses a frame.  But it also adds a generic check
to catch reusage without resetting the flags, via a warning bit set on
exit from Do_Core