Fix CHAR! add/subtract from CHAR, add enfix tests
R3-Alpha changed the behavior of math on CHAR! when subtracting another
CHAR!, giving an INTEGER! of the delta.  (Rebol2 would return a CHAR!
with the codepoint of the difference, overflowing in the 8-bit space.
Red gives a CHAR! but errors on overflow).

The behavior had become broken and wasn't giving signed differences,
but was unnoticed because the character tests hadn't been modified for
R3-Alpha.  This fixes the behavior and updates the Rebol2 tests.

Also adds enfix tests, to mirror the examples given in the Trello.