Triage and delete %r2-tests.r
The original R3-Alpha test suite was designed to compare and contrast
tests with Rebol2, and could also be run with Rebol2.  That was deemed
unimportant when integrating the tests into Ren-C and trying to
formalize them, so those tests which would only pass in Rebol2 were
removed from the suite.

Rather than delete them, they were collected into a file to be looked
at later.  Many of the tests had parallel versions that had been
adapted for R3-Alpha, but not all did.  Some had yet to have the tests
converted to adapt to the changes, some were for features that had not
been implemented, some were irrelevant, and some were just placeholders
for issues that hadn't been resolved.

This commit includes what triage could be done for those issues in
terms of making new tests.  Github issues or Trello cards were made
for the other issues, just to get them out of the code and into a
place where they might be discussed.  The upshot is the deletion of
the %tests/old/r2-tests.r file, with all information believed to be
extracted from it to more actionable locations.