Respect array index in BIND without /COPY, clipboard tests
At one point it was noticed that BIND was only paying attention to the
index of the passed in ANY-ARRAY! if it was copied.  A comment to this
effect was added, but it wasn't changed for fear of breaking something
that depended on that behavior.

However it turns out that there was an actual bug already raised about
the issue, and a breaking test based on it:

https://github.com/rebol/rebol-issues/issues/2086

Since that test is probably a very rare example of code that depends
on the behavior, and it's fixed by changing it, this goes ahead and
bites the bullet and fixes it.

Also stubs out a directory for the clipboard module as a place to hold
tests for it, and updates those tests so they pass for the existing
Windows clipboard implementation.