Fix PNG decoding/encoding, update LodePNG, resurrect tests
Some PNG encoding and decoding tests had been removed from the main
Ren-C tests at some point.  Possibly because they didn't work after
merging in the Saphir/Atronix image codec changes.  (They worked
in R3-Alpha.)  Or perhaps because they were part of a group where
some tests failed because of missing codecs (JPG/GIF encoding).

Whatever the reason, the absence of these tests meant some checking on
round-tripping and equivalencies of lossless encodings was not being
done.  Reinstating those tests discovered that although the PNG
decode and encode process was able to round-trip a file that would
load correctly, the actual bytes in-memory for the image were not
consistent with what BMP or GIF decoding would produce.  (So if there
were an image viewer within Rebol tested, it would have shown wrong
colors for the PNG.)

As part of looking into the problem, this went ahead and synced up
LodePNG to the current version from GitHub.

However, the real issue is that IMAGE! uses "quad" values to access
pixels, as opposed to going according to bytes.  This means that it
has to be sensitive to the byte ordering of the underlying platform,
and being compatible with it wasn't being done correctly.

The proabable "right" answer to dealing with this is to standardize
the byte order underlying an image to RGBA, and do that access via
individual unsigned-char access, so that the endianness is not an
issue.  If a BINARY! series were used to do this, it would also mean
that rebRepossess() could be used to avoid creating an secondary
series for the image distinct from what the PNG decoding uses.

That is a separate issue, however.  This fixes the code for the moment,
and makes the tests pass.  It also adds a usage of rebRepossess()
in the encoder, rather than copy a series that doesn't need to be
copied.