Fix DEHEX behavior w.r.t. UTF-8, add corresponding ENHEX
R3-Alpha did not follow RFC 3896, which states that percent-encoded
URLs should be interpreting the percent encoded data as UTF-8.  This
was a longstanding bug report:

https://github.com/rebol/rebol-issues/issues/1986

A test was added that would fail, since the issue had not been fixed.
This test (along with many others) was put aside for later review
when the tests were changed to require all tests pass.

As part of an effort to fully triage those "put aside" tests (in one
way or another), this just goes ahead and fixes the issue.  It also
adds a corresponding ENHEX native which follows the RFC for producing
hex bytes for characters that need it.  Per that RFC, all hex byte
characters are canonized to uppercase in production (though both
upper and lower are tolerated by dehex).

One change in behavior is that a % sign in the source material *must*
be followed by two hex digits, either upper or lower case.  Other
appearances of % will be treated as an error.