Use comparison of start/end to direct limit test in FOR
Issue #1993 boils down to a fairly simple question of how the R2-styled
FOR loop decides if it is counting upward or downward.  Historically
it used the sign of the `bump` variable, so if it was positive it
would check to see if the state was greater than or equal to the end
value...and if it was negative it would check to see if the state was
less than or equal to the end value.

However, this meant that a bump value of 0 would either have to be made
an error, or have a seemingly asymmetric behavior resulting from
deciding that zero was either positive or negative.

Rather than making 0 illegal, it was proposed that the relative values
of the start and the end be what controls the direction of comparison.
This implements the proposal.

The ticket discussing the proposal turned from something fairly mundane
into something very lengthy, which has been moderated somewhat in order
to hopefully make the issue and arguments more approachable:

https://github.com/rebol/rebol-issues/issues/1993

This allowed some tests which had been moved aside to be brought back
into the suite.