Run body once if start=end in loops
An oversight in the previous commit meant it wasn't running loops
the single time it should, when start=end...it was returning void.

Since the loops don't need to do an increment if start=end, then no
such cases overflow, so this affected tests as well.

The tests caught this, but the commit was pushed in an oversight.