Allow BIND to SELF
Although SELF is a hidden field, it is also considered "bindable".
This concept is somewhat nebulous, but R3-Alpha had the idea of these
two things being different...though it applied only to SELF (while
now it's a generic bit that could, theoretically, be applied to
other things).

It may be that no hidden fields are bindable from outside the object,
but for the moment this changes it to be compatible with R3-Alpha,
as the bit does exist...so it makes it clearer, for now.

Includes some improvements to comments.