Triage and categorize tests in %pending-tests.r, remove file
There was a practice in the R3-Alpha tests of writing and committing
tests for things that did not work, and then "diffing" the test results
against a version of the tests run using another version.  This unusual
practice was not preserved in Ren-C, which instead sought to get a
clean run of the tests with 0 failures...that could be depended on and
interpreted using the version of the interpreter those tests were
executed with.

Getting to that point required several phases, one of fixing all places
where test code could actually *crash* the interpreter so it was--at
worst--an error.  The next phase was to fix as many bugs as possible,
though some were moved to a %pending-tests.r file for later review as
looking into them would take additional time.

In the interests of not having a "test dumpster" sitting around to trip
over when doing searches in the code, this is the result of going
through all the non-working tests and either fixing the relevant issue,
or making sure the test code is properly attached to an issue in the
bug database to be reviewed.  The result is being able to eliminate
the %pending-tests.r file.