Give HOST-CONSOLE second chance on typos/errors
It's not terribly likely that an error or throw from the HOST-CONSOLE
function can be salvaged just by trying to run it again.  But giving
it a second chance instead of crashing outright could be helpful in
some debugging scenarios.  This gives it one chance to recover, and
resets that chance once it successfully requests an evaluation on
behalf of the user.