Replace LOGICAL() with DID() as complement to NOT()
Not until C99 with a true boolean type was it possible to guarantee a
`bool` parameter would always be coerced to a 0 or a 1.  REBOOL was
just an alias for an integer type, so it could wind up holding other
values if one was not careful.

The LOGICAL() and NOT() macros were used to coerce arbitrary integer
expressions into REBOOLs, and some trickery was used to do compilation
in a mode that could check that they were always used.  More recently,
Rebol adopted the DID keyword as the complement to NOT, which is
shorter and makes about as much sense (why would `not "hello"` be
false, anyway?)

This embraces DID as a replacement for LOGICAL.  It also brings up to
date the "strict bool" test, by making it C++11 only, so that it will
actually produce a working executable.  (This test should be added
into one of the Travis build matrix entries...)