Drop experimental DO-ALL native (for now)
DO-ALL was a concept which would run multiple clauses separated by
BAR! and propagate failure while still running all the clauses.  To
do this, it would try to "recover" a frame out of an erroring state.

This was rather ad-hoc, and was never intended to work with C va_list
frames, which means that API calls using DO-ALL would not be able to
work as written.  This meant Recover_Frame() was not very robust, and
this is compounded with a previous finding that the workings with
throws wasn't quite right either.

The best move for the moment is probably to remove the routine...
in order to focus on more pressing issues than something with complex
semantics and no callers.