Remove DID from PARSE tests, add PARSE-MATCH "enclosing"
There was a proposal here that would change the result type of PARSE:

https://github.com/rebol/rebol-issues/issues/2165

It was tried for a time, but turned out to make some of the simpler and
more common PARSE cases seem more overwhelming with non-trivial input.

This removes DID from the %parse-tests.r since it was basically just
more comfortable to return a LOGIC!.  However, it also adds PARSE-MATCH
to show how Ren-C's ENCLOSE could be used to easily create a variant of
PARSE which has the desired behavior.