Correct strict aliasing mistakes using uint8_t
On most systems, uint8_t is defined as unsigned char.  However, in the
rare case where it is *not* defined thusly, it should not be used as
a generic "byte" type to access memory at a pointer to another type.
This is because only `char*` and `unsigned char*` are exempt from the
"strict aliasing rules":

https://stackoverflow.com/q/37259909/
https://stackoverflow.com/a/16138470/

Given how REBYTE is used, this means it certainly should be defined as
unsigned char, not uint8_t.  It also means other locations that are
reading memory that was written by other types should use either char
or unsigned char to do so.

This reverts a very recent change to the definition of REBYTE, and
other locations.