Resurrect `<r3-legacy>` emulation, via more usermode code
The `do <r3-legacy>` compatibility mode had atrophied somewhat from
lack of use, and several of the code paths inside the interpreter's C
code were removed to tighten things up.

This brings back a less-bend-over-backward compatibility mode, that is
not geared toward emulating obscure quirks of R3-Alpha.  Rather it acts
in the interest of aiming to enable a stylization of code as a "lowest
common denominator" which could run in R3-Alpha/Rebol2/Red, or this
`<r3-legacy>` mode of Ren-C.

To accomplish this, it does things like provide augmented FUNC and
FUNCTION generators that use FRAME! reflection to translate LOGIC!
false refinements into BLANK! values, and void arguments to unused
refinements into BLANK!s.  These function generators also provide a
definitional EXIT, in the spirit of the usermode form of definitional
returns...rather than try to directly simulate Red/Rebol2/R3-Alpha's
"climb the stack and exit the first non-native function" behavior
(which was bad).

This change was enough to run the Draem blog code from 2014, with
minor modifications, most all of which were related to the little-used
/ONLY feature for "don't run blocks" on conditionals, and the ability
to use non-blocks in conditional slots.  Explained here:

https://github.com/hostilefork/draem/commit/e9720a56ffaab9eec69a70ece76e66a43245b597