Add ++, --, COMMENT to `<r3-legacy>`; fix legacy HAS
Brings `<r3-legacy>` the prefix quoting ++ and -- constructs from
R3-Alpha, which were labeled "new, hackish stuff" in comments.

This makes COMMENT return a void, instead of disappearing completely.
Probably not much code depended on that, but as the legacy acts as
a sort of running narration on compatibility, it doesn't hurt.

Also, the HAS function as written would not support EXIT or the
blanking conventions of the locals.  This factors out the R3-ALPHA-FUNC
so it can be used in the legacy HAS.