Adjust window position on Windows
Due to the mismatch between Rebol and SDL, on windows, a window opens at
(0, 0) will not show the title bar, because it tries to position the
top-left corner of the client aren at (0, 0), which pushs the window
title bar out of the screen. Just offseting by the title hight makes it
show up.