Implement graphic cursor manipulation
This is ugly, because in AGG, it just blindly passes the cursor ID to
the underlying system functions, so each platform has different IDs. ZOE
is using different IDs for different platforms. However, SDL has a
common ID for each cursor, so it needs to converge the different IDs
from user script and feed it to SDL, which will then be mapped to
platform specific ID.