Turn off touch event on Windows
Because Rebol/ZOE doesn't support touch event differently, they have to
be converted to mouse events. Turning off the touch event will shift the
conversion work from the application to the OS, and the result is much
better. e.g. press and hold will be mapped to a right click, etc.